Invert dependencies to providers
I added a injector widget. This widget takes an injector. The injector widget is the parent of all other widgets in the app. This way any widget can access the injector widget. From the injector widget, get the injector, which contains all dependencies. For testing, just pass a different injector to the injector widget.